- update readme - rename filter_page to favourite_cocktails_screen
