+++ lesson_14/homework/README.md
@@ -1,16 +1,5 @@
-# homework
-
-A new Flutter project.
-
-## Getting Started
-
-This project is a starting point for a Flutter application.
-
-A few resources to get you started if this is your first Flutter project:
-
-- [Lab: Write your first Flutter app](https://flutter.dev/docs/get-started/codelab)
-- [Cookbook: Useful Flutter samples](https://flutter.dev/docs/cookbook)
-
-For help getting started with Flutter, view our
-[online documentation](https://flutter.dev/docs), which offers tutorials,
-samples, guidance on mobile development, and a full API reference.
+- Открыть класс экрана FavouriteCocktailsScreen
+- Внести изменения в код экрана, выделив соотвествующие обработчики событий на нажатие элемента из списка избранных коктейлей
+- В обработчике события на переход к экрану описания коктейля выполнить асинхронное получение модели коктейля
+- В обработчике события на переход к экрану описания коктейля после получения модели коктейля выполнить навигацию к экрану описания коктейля (CocktailDetailsScreen) с передачей полученной модели
+- Убедиться, что CocktailDetailsScreen отображает корректно полученную модель
